Benchmark misread.qc function with dichotomized-Gaussian simulations

This function takes the same summary statistics (STdata) and barcode codebook (codebook) as misread.qc and runs dichotomized-Gaussian simulations with stipulated bit-flip rates and bit-flip correlations in order to estimate how well the L-BFGS algorithm recovers the bit-flip rates and bit-flip correlations for the given data set and codebook.

Usage

dichot.guass.benchmark(
  STdata,
  codebook,
  n_sims = 100,
  n_forks = 1,
  max_flips = 0,
  report_freq = 10,
  maxeval = 500,
  max_correctable_Hamming_distance = NULL
)

Arguments

STdata

Numeric matrix with rows as barcodes, columns labeled "rates", "variance", "counts". Must have barcode names (e.g., gene or protein species) as row names.

codebook

Numeric matrix with barcodes as row names and bits as columns. All entries should be 1 or 0, depending on whether an mRNA molecule of the species represented by the row is expected to luminescence in the bit represented by the column. All row names from STdata must be included in the row names for codebook.

n_sims

Number of dichotomized-Gaussian simulations to run. The default is 100.

n_forks

Number of process forks to use for expected-count computation (i.e., parallel computation), default is 1. Must be 1 on Windows, can be higher on Mac and Linux.

max_flips

When analytically computing expected corrected counts per barcode, the function will ignore misreads larger than this Hamming distance. The default is 0, which is interpreted as no limit. Using all misreads will likely be prohibitively slow; a value between six and ten is probably advisable. Values of 3 or 4 work well for initial trouble shooting and testing.

report_freq

Divisor specifying report frequency during optimization; will print updates every report_freq accepted calls, default 10.

maxeval

Maximum number of objective function evaluations for L-BFGS, default 500.

max_correctable_Hamming_distance

Maximum Hamming distance for misreads to be corrected, default NULL sets it to one less than the minimum Hamming distance between codebook entries.

Value

A list giving:

fr_est: A matrix giving the flip rates and bit-flip correlations estimated by misread.qc (columns), for each set of observed counts generated by dichotomized-Gaussian simulation (rows).

fr_stipulated: A vector giving the stipulated flip rates and bit-flip correlations used to generate the dichotomized-Gaussian simulations. All simulations use the same stipulated values.

PPV_est: A matrix giving positive predictive value (PPV) expected based on analytical computation, using the values in fr_est, per barcode (columns), for each simulation run (rows).

PPV_expected: A vector giving the PPV values expected based on analytical computation for each barcode, given the stipulated flip rates and stipulated bit-flip correlations in fr_stipulated.

ecc_est: A matrix giving the corrected count expected based on analytical computation, using the values in fr_est, per barcode (columns), for each simulation run (rows).

ecc_expected: A vector giving the corrected count expected based on analytical computation for each barcode, given the stipulated flip rates and stipulated bit-flip correlations in fr_stipulated.

sim_counts: A matrix giving the actual simulated count per barcode (columns) for each simulation (rows).